AutoTrust/django-AutoTrust/requirements.txt
# Framework principal
Django==4.2.16

# Herramientas útiles de desarrollo
django-environ==0.11.2  # Para manejar variables de entorno (.env)
djangorestframework==3.15.2  # Para crear APIs (REST)

# Seguridad y utilidades
python-decouple==3.8    # Alternativa ligera para leer configuraciones
Pillow==11.0.0          # Para manejar imágenes en Django
